+++ packages/shared/src/lib/core/market/actions/index.ts
@@ -1,3 +1,5 @@
 export * from './getAndUpdateMarketPrices'
+export * from './getMarketAmountFromTokenValue'
+export * from './getMarketPriceForToken'
 export * from './getMarketPrices'
 export * from './market-prices-polling'
