--- packages/shared/src/lib/core/market/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './getMarketAmountFromTokenValue'
-export * from './getMarketPriceForToken'
